🚀 feat: 詳細画面の仮実装

+++ lib/data/datasources/remote/firebase_item_datasource.dart
@@ -18,6 +18,20 @@ class FirebaseItemDataSource {
         .get();
   }
 
+  Future<QuerySnapshot> retrieveTimelineItems(String itemId) {
+    final user = _auth.currentUser;
+    if (user == null) {
+      throw Exception('User is not logged in');
+    }
+    return _firestore
+        .collection('users')
+        .doc(user.uid)
+        .collection('lists')
+        .doc(itemId)
+        .collection('timeline')
+        .get();
+  }
+
   Future<DocumentReference> createItem(Map<String, dynamic> item) {
     final user = _auth.currentUser;
     if (user == null) {
@@ -28,6 +42,23 @@ class FirebaseItemDataSource {
         .doc(user.uid)
         .collection('lists')
         .add(item);
+  }
+
+  Future<DocumentReference> createTimelineItem(
+    String itemId,
+    Map<String, dynamic> timelineItem,
+  ) {
+    final user = _auth.currentUser;
+    if (user == null) {
+      throw Exception('User is not logged in');
+    }
+    return _firestore
+        .collection('users')
+        .doc(user.uid)
+        .collection('lists')
+        .doc(itemId)
+        .collection('timeline')
+        .add(timelineItem);
   }
 
   Future<void> updateItem(String id, Map<String, dynamic> item) {
@@ -43,6 +74,25 @@ class FirebaseItemDataSource {
         .update(item);
   }
 
+  Future<void> updateTimelineItem(
+    String itemId,
+    String timelineItemId,
+    Map<String, dynamic> item,
+  ) {
+    final user = _auth.currentUser;
+    if (user == null) {
+      throw Exception('User is not logged in');
+    }
+    return _firestore
+        .collection('users')
+        .doc(user.uid)
+        .collection('lists')
+        .doc(itemId)
+        .collection('timeline')
+        .doc(timelineItemId)
+        .update(item);
+  }
+
   Future<void> deleteItem(String id) {
     final user = _auth.currentUser;
     if (user == null) {
@@ -53,6 +103,21 @@ class FirebaseItemDataSource {
         .doc(user.uid)
         .collection('lists')
         .doc(id)
+        .delete();
+  }
+
+  Future<void> deleteTimelineItem(String itemId, String timelineItemId) {
+    final user = _auth.currentUser;
+    if (user == null) {
+      throw Exception('User is not logged in');
+    }
+    return _firestore
+        .collection('users')
+        .doc(user.uid)
+        .collection('lists')
+        .doc(itemId)
+        .collection('timeline')
+        .doc(timelineItemId)
         .delete();
   }
 }
